refactor single file into logical multifiles
