+++ .gitignore
@@ -72,3 +72,7 @@ target/
 
 # pyenv
 .python-version
+
+# Editors
+.idea/
+.vscode/
