+++ setup.cfg
@@ -41,4 +41,4 @@ where=src
 
 [options.entry_points]
 pytest11 =
-    rich = pytest_rich
+    rich = pytest_rich.plugin
